[#89] feat: 초대하기 모달 대시보드 수정 페이지 연결
